Allow errors to be ignored by the builder
