+++ pkg/kubectl/resource/builder.go
@@ -21,16 +21,13 @@ import (
 	"io"
 	"net/url"
 	"os"
-	"reflect"
 	"strings"
 
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/api"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/api/meta"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/labels"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/runtime"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util"
 	"github.com/GoogleCloudPlatform/kubernetes/pkg/util/errors"
-	"github.com/GoogleCloudPlatform/kubernetes/pkg/watch"
 )
 
 // Builder provides convenience functions for taking arguments and parameters
@@ -406,156 +403,6 @@ func (b *Builder) Do() *Result {
 	}
 	r.visitor = NewDecoratedVisitor(r.visitor, helpers...)
 	return r
-}
-
-// Result contains helper methods for dealing with the outcome of a Builder.
-type Result struct {
-	err     error
-	visitor Visitor
-
-	sources  []Visitor
-	singular bool
-
-	// populated by a call to Infos
-	info []*Info
-}
-
-// Err returns one or more errors (via a util.ErrorList) that occurred prior
-// to visiting the elements in the visitor. To see all errors including those
-// that occur during visitation, invoke Infos().
-func (r *Result) Err() error {
-	return r.err
-}
-
-// Visit implements the Visitor interface on the items described in the Builder.
-// Note that some visitor sources are not traversable more than once, or may
-// return different results.  If you wish to operate on the same set of resources
-// multiple times, use the Infos() method.
-func (r *Result) Visit(fn VisitorFunc) error {
-	if r.err != nil {
-		return r.err
-	}
-	return r.visitor.Visit(fn)
-}
-
-// IntoSingular sets the provided boolean pointer to true if the Builder input
-// reflected a single item, or multiple.
-func (r *Result) IntoSingular(b *bool) *Result {
-	*b = r.singular
-	return r
-}
-
-// Infos returns an array of all of the resource infos retrieved via traversal.
-// Will attempt to traverse the entire set of visitors only once, and will return
-// a cached list on subsequent calls.
-func (r *Result) Infos() ([]*Info, error) {
-	if r.err != nil {
-		return nil, r.err
-	}
-	if r.info != nil {
-		return r.info, nil
-	}
-	infos := []*Info{}
-	err := r.visitor.Visit(func(info *Info) error {
-		infos = append(infos, info)
-		return nil
-	})
-	r.info, r.err = infos, err
-	return infos, err
-}
-
-// Object returns a single object representing the output of a single visit to all
-// found resources.  If the Builder was a singular context (expected to return a
-// single resource by user input) and only a single resource was found, the resource
-// will be returned as is.  Otherwise, the returned resources will be part of an
-// api.List. The ResourceVersion of the api.List will be set only if it is identical
-// across all infos returned.
-func (r *Result) Object() (runtime.Object, error) {
-	infos, err := r.Infos()
-	if err != nil {
-		return nil, err
-	}
-
-	versions := util.StringSet{}
-	objects := []runtime.Object{}
-	for _, info := range infos {
-		if info.Object != nil {
-			objects = append(objects, info.Object)
-			versions.Insert(info.ResourceVersion)
-		}
-	}
-
-	if len(objects) == 1 {
-		if r.singular {
-			return objects[0], nil
-		}
-		// if the item is a list already, don't create another list
-		if _, err := runtime.GetItemsPtr(objects[0]); err == nil {
-			return objects[0], nil
-		}
-	}
-
-	version := ""
-	if len(versions) == 1 {
-		version = versions.List()[0]
-	}
-	return &api.List{
-		ListMeta: api.ListMeta{
-			ResourceVersion: version,
-		},
-		Items: objects,
-	}, err
-}
-
-// ResourceMapping returns a single meta.RESTMapping representing the
-// resources located by the builder, or an error if more than one
-// mapping was found.
-func (r *Result) ResourceMapping() (*meta.RESTMapping, error) {
-	if r.err != nil {
-		return nil, r.err
-	}
-	mappings := map[string]*meta.RESTMapping{}
-	for i := range r.sources {
-		m, ok := r.sources[i].(ResourceMapping)
-		if !ok {
-			return nil, fmt.Errorf("a resource mapping could not be loaded from %v", reflect.TypeOf(r.sources[i]))
-		}
-		mapping := m.ResourceMapping()
-		mappings[mapping.Resource] = mapping
-	}
-	if len(mappings) != 1 {
-		return nil, fmt.Errorf("expected only a single resource type")
-	}
-	for _, mapping := range mappings {
-		return mapping, nil
-	}
-	return nil, nil
-}
-
-// Watch retrieves changes that occur on the server to the specified resource.
-// It currently supports watching a single source - if the resource source
-// (selectors or pure types) can be watched, they will be, otherwise the list
-// will be visited (equivalent to the Infos() call) and if there is a single
-// resource present, it will be watched, otherwise an error will be returned.
-func (r *Result) Watch(resourceVersion string) (watch.Interface, error) {
-	if r.err != nil {
-		return nil, r.err
-	}
-	if len(r.sources) != 1 {
-		return nil, fmt.Errorf("you may only watch a single resource or type of resource at a time")
-	}
-	w, ok := r.sources[0].(Watchable)
-	if !ok {
-		info, err := r.Infos()
-		if err != nil {
-			return nil, err
-		}
-		if len(info) != 1 {
-			return nil, fmt.Errorf("watch is only supported on a single resource - %d resources were found", len(info))
-		}
-		return info[0].Watch(resourceVersion)
-	}
-	return w.Watch(resourceVersion)
 }
 
 func SplitResourceArgument(arg string) []string {
